feature : #74

+++ Jenkinsfile
@@ -47,8 +47,15 @@ pipeline {
                         bat "docker login -u %DOCKERHUB_USERNAME% -p %DOCKERHUB_PASSWORD%"
 
                         // 기존 컨테이너를 중지하고 제거
-                        bat "docker stop ${dockerContainerName} > nul 2>&1 || ( echo Container not running or does not exist. )"
-                        bat "docker rm ${dockerContainerName} > nul 2>&1 || ( echo Container not running or does not exist. )"
+                        def dockerPsOutput = bat(script: "docker ps -a --filter name=${dockerContainerName}", returnStdout: true).trim()
+
+                           if (dockerPsOutput.contains(dockerContainerName)) {
+                               // 기존 컨테이너를 중지하고 제거
+                               bat "docker stop ${dockerContainerName} > nul 2>&1 || ( echo Container not running or does not exist. )"
+                               bat "docker rm ${dockerContainerName} > nul 2>&1 || ( echo Container not running or does not exist. )"
+                           } else {
+                               echo "컨테이너가 없습니다."
+                           }
 
                         // DockerHub에 생성한 이미지 push
                         bat "docker push ${dockerImageName}"
